+++ test/parallel_api/dynamic_selection/sycl/test_auto_tune_policy_sycl.pass.cpp
@@ -13,7 +13,6 @@
 #include "support/test_dynamic_selection_utils.h"
 #include "support/test_config.h"
 #if TEST_DYNAMIC_SELECTION_AVAILABLE
-#    include "support/sycl_sanity.h"
 
 int
 test_auto_initialization(const std::vector<sycl::queue>& u)
@@ -108,11 +107,11 @@ test_auto_submit_wait_on_event(UniverseContainer u, int best_resource)
                     }
                 }
                 ecount += i;
-                if (*j == 0) 
+                if (*j == 0)
                 {
                      return sycl::event{};
-                } 
-                else 
+                }
+                else
                 {
                     return q.submit([=](sycl::handler& h) {
                         h.parallel_for<TestUtils::unique_kernel_name<
@@ -153,11 +152,11 @@ test_auto_submit_wait_on_event(UniverseContainer u, int best_resource)
                         }
                     }
                     ecount += i;
-                    if (*j == 0) 
+                    if (*j == 0)
                     {
                          return sycl::event{};
-                    } 
-                    else 
+                    }
+                    else
                     {
                         return q.submit([=](sycl::handler& h) {
                             h.parallel_for<TestUtils::unique_kernel_name<
@@ -250,11 +249,11 @@ test_auto_submit_wait_on_group(UniverseContainer u, int best_resource)
                     }
                 }
                 ecount += i;
-                if (*j == 0) 
+                if (*j == 0)
                 {
                      return sycl::event{};
-                } 
-                else 
+                }
+                else
                 {
                     return q.submit([=](sycl::handler& h) {
                         h.parallel_for<TestUtils::unique_kernel_name<
@@ -295,11 +294,11 @@ test_auto_submit_wait_on_group(UniverseContainer u, int best_resource)
                         }
                     }
                     ecount += i;
-                    if (*j == 0) 
+                    if (*j == 0)
                     {
                          return sycl::event{};
-                    } 
-                    else 
+                    }
+                    else
                     {
                         return q.submit([=](sycl::handler& h) {
                             h.parallel_for<TestUtils::unique_kernel_name<
@@ -392,11 +391,11 @@ test_auto_submit_and_wait(UniverseContainer u, int best_resource)
                     }
                 }
                 ecount += i;
-                if (*j == 0) 
+                if (*j == 0)
                 {
                      return sycl::event{};
-                } 
-                else 
+                }
+                else
                 {
                     return q.submit([=](sycl::handler& h) {
                         h.parallel_for<TestUtils::unique_kernel_name<
@@ -436,11 +435,11 @@ test_auto_submit_and_wait(UniverseContainer u, int best_resource)
                         }
                     }
                     ecount += i;
-                    if (*j == 0) 
+                    if (*j == 0)
                     {
                          return sycl::event{};
-                    } 
-                    else 
+                    }
+                    else
                     {
                         return q.submit([=](sycl::handler& h) {
                             h.parallel_for<TestUtils::unique_kernel_name<
@@ -486,7 +485,6 @@ build_auto_tune_universe(std::vector<sycl::queue>& u)
     {
         auto device_cpu1 = sycl::device(sycl::cpu_selector_v);
         sycl::queue cpu1_queue(device_cpu1);
-        run_sycl_sanity_test(cpu1_queue);
         u.push_back(cpu1_queue);
     }
     catch (const sycl::exception&)
@@ -497,7 +495,6 @@ build_auto_tune_universe(std::vector<sycl::queue>& u)
     {
         auto device_cpu2 = sycl::device(sycl::cpu_selector_v);
         sycl::queue cpu2_queue(device_cpu2);
-        run_sycl_sanity_test(cpu2_queue);
         u.push_back(cpu2_queue);
     }
     catch (const sycl::exception&)
@@ -508,7 +505,6 @@ build_auto_tune_universe(std::vector<sycl::queue>& u)
     {
         auto device_cpu3 = sycl::device(sycl::cpu_selector_v);
         sycl::queue cpu3_queue(device_cpu3);
-        run_sycl_sanity_test(cpu3_queue);
         u.push_back(cpu3_queue);
     }
     catch (const sycl::exception&)
@@ -519,7 +515,6 @@ build_auto_tune_universe(std::vector<sycl::queue>& u)
     {
         auto device_cpu4 = sycl::device(sycl::cpu_selector_v);
         sycl::queue cpu4_queue(device_cpu4);
-        run_sycl_sanity_test(cpu4_queue);
         u.push_back(cpu4_queue);
     }
     catch (const sycl::exception&)
