+++ test/parallel_api/dynamic_selection/sycl/test_dynamic_load_policy_sycl.pass.cpp
@@ -12,7 +12,6 @@
 #include "support/test_dynamic_load_utils.h"
 #include "support/test_config.h"
 #if TEST_DYNAMIC_SELECTION_AVAILABLE
-#    include "support/sycl_sanity.h"
 
 static inline void
 build_dl_universe(std::vector<sycl::queue>& u)
@@ -21,7 +20,6 @@ build_dl_universe(std::vector<sycl::queue>& u)
     {
         auto device_cpu1 = sycl::device(sycl::cpu_selector_v);
         sycl::queue cpu1_queue(device_cpu1);
-        run_sycl_sanity_test(cpu1_queue);
         u.push_back(cpu1_queue);
     }
     catch (const sycl::exception&)
@@ -32,7 +30,6 @@ build_dl_universe(std::vector<sycl::queue>& u)
     {
         auto device_cpu2 = sycl::device(sycl::cpu_selector_v);
         sycl::queue cpu2_queue(device_cpu2);
-        run_sycl_sanity_test(cpu2_queue);
         u.push_back(cpu2_queue);
     }
     catch (const sycl::exception&)
