+++ test/support/test_dynamic_selection_utils.h
@@ -15,6 +15,46 @@
 #include <random>
 #include <algorithm>
 
+namespace TestUtils
+{
+  template <sycl::usm::alloc alloc_type>
+  constexpr ::std::size_t
+  uniq_kernel_index()
+  {
+    return static_cast<typename ::std::underlying_type_t<sycl::usm::alloc>>(alloc_type);
+  }
+
+  template <typename Op, ::std::size_t CallNumber>
+  struct unique_kernel_name;
+}
+
+static inline void
+build_universe(std::vector<sycl::queue> &u) {
+  try {
+    auto device_default = sycl::device(sycl::default_selector_v);
+    sycl::queue default_queue(device_default);
+    u.push_back(default_queue);
+  } catch (const sycl::exception&) {
+    std::cout << "SKIPPED: Unable to run with default_selector\n";
+  }
+
+  try {
+    auto device_gpu = sycl::device(sycl::gpu_selector_v);
+    sycl::queue gpu_queue(device_gpu);
+    u.push_back(gpu_queue);
+  } catch (const sycl::exception&) {
+    std::cout << "SKIPPED: Unable to run with gpu_selector\n";
+  }
+
+  try {
+    auto device_cpu = sycl::device(sycl::cpu_selector_v);
+    sycl::queue cpu_queue(device_cpu);
+    u.push_back(cpu_queue);
+  } catch (const sycl::exception&) {
+    std::cout << "SKIPPED: Unable to run with cpu_selector\n";
+  }
+}
+
 template <typename Policy, typename T>
 int
 test_initialization(const std::vector<T>& u)
